A beta version for tree Genetic Programming

+++ libeasea/include/CEvolutionaryAlgorithm.h
@@ -59,8 +59,8 @@ public:
   Parameters* params;
 
   CGnuplot* gnuplot;
-  float currentAverageFitness;
-  float currentSTDEV;
+  double currentAverageFitness;
+  double currentSTDEV;
 
   virtual ~CEvolutionaryAlgorithm();
